Implement case list view Change navigation graph for login to case fragment Temporarily disable rotation to avoid configuration changes Improve logic for optional navigation Implement date and missing persons name to list items

+++ app/src/main/java/com/example/sarsolutions/models/DTO.kt
@@ -18,7 +18,7 @@ data class Case(
     val missingPersonName: List<String>,
     val reporterName: String,
     val volunteers: List<Volunteer>,
-    val date: Int
+    val date: Long
 )
 
 data class Volunteer(
